+++ www/src/components/plugin-searchbar-body.js
@@ -13,8 +13,7 @@ import { Link } from "gatsby"
 import DownloadArrow from "react-icons/lib/md/file-download"
 import AlgoliaLogo from "../assets/algolia.svg"
 import GatsbyIcon from "../monogram.svg"
-import debounce from "lodash/debounce"
-import unescape from "lodash/unescape"
+import { debounce, unescape } from "lodash-es"
 
 import {
   space,
@@ -22,7 +21,7 @@ import {
   fontSizes,
   transition,
   radii,
-  breakpoints,
+  mediaQueries,
   sizes,
   fonts,
 } from "../utils/presets"
@@ -197,7 +196,7 @@ class Search extends Component {
       <div
         css={{
           paddingBottom: rhythm(2.5),
-          [breakpoints.md]: {
+          [mediaQueries.md]: {
             paddingBottom: 0,
           },
         }}
@@ -260,7 +259,7 @@ class Search extends Component {
         <div>
           <div
             css={{
-              [breakpoints.md]: {
+              [mediaQueries.md]: {
                 height: `calc(100vh - ${sizes.headerHeight} - ${
                   sizes.bannerHeight
                 } - ${searchInputHeight} - ${searchInputWrapperMargin} - ${searchMetaHeight})`,
@@ -465,7 +464,10 @@ class PluginSearchBar extends Component {
   urlToSearch = () => {
     if (this.props.location.search) {
       // ignore this automatically added query parameter
-      return this.props.location.search.replace(`no-cache=1`, ``).slice(2)
+      const search = this.props.location.search
+        .replace(`no-cache=1`, ``)
+        .slice(2)
+      return decodeURIComponent(search)
     }
     return ``
   }