+++ www/src/pages/contributing/index.js
@@ -7,6 +7,7 @@ import { itemListContributing } from "../../utils/sidebar/item-list"
 import Container from "../../components/container"
 import EmailCaptureForm from "../../components/email-capture-form"
 import DocSearchContent from "../../components/docsearch-content"
+import FooterLinks from "../../components/shared/footer-links"
 
 class IndexRoute extends React.Component {
   render() {
@@ -63,6 +64,8 @@ class IndexRoute extends React.Component {
               </li>
             </ul>
             <EmailCaptureForm signupMessage="Want to keep up with the latest tips &amp; tricks? Subscribe to our newsletter!" />
+
+            <FooterLinks />
           </Container>
         </DocSearchContent>
       </Layout>