+++ www/src/views/starter-library/starter-list.js
@@ -8,7 +8,7 @@ import styles from "../shared/styles"
 import ThumbnailLink from "../shared/thumbnail"
 import EmptyGridItems from "../shared/empty-grid-items"
 import V2Icon from "../../assets/v2icon.svg"
-import get from "lodash/get"
+import { get } from "lodash-es"
 
 const StartersList = ({ urlState, starters, count, sortRecent }) => {
   if (!starters.length) {
@@ -117,10 +117,6 @@ const StartersList = ({ urlState, starters, count, sortRecent }) => {
                         <strong className="title">{name}</strong>
                       </h5>
                     </Link>
-                    {/* {isGatsbyVersionWarning ?
-                        <span css={{ fontStyle: `italic`, color: `red` }}>Outdated Version: {minorVersion}</span> :
-                        <span css={{ fontStyle: `italic`, color: `green` }}>Gatsby Version: {minorVersion}</span>
-                      } */}
                   </div>
                   <div
                     css={{
@@ -178,7 +174,7 @@ const StartersList = ({ urlState, starters, count, sortRecent }) => {
 
 export default StartersList
 
-function sortingFunction(sortRecent) {
+function sortingFunction() {
   return function({ node: nodeA }, { node: nodeB }) {
     const metricA = get(nodeA, `fields.starterShowcase.stars`, 0)
     const metricB = get(nodeB, `fields.starterShowcase.stars`, 0)